Fixing build events. Removing warnings. Implementing glad debugger on every draw call. Also refactoring the obj reader slightly.

+++ Source/ProspectEngine/Libraries/MaterialLibrary_impl.cpp
@@ -105,5 +105,5 @@ Material MaterialLibrary_impl::AddMaterial(
 
 int MaterialLibrary_impl::GetMaterialCount() const
 {
-   return m_materials.size();
+   return static_cast<int>(m_materials.size());
 }
